+++ lib/map_state_to_deployments.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/opentable/sous/util/logging"
+	"github.com/opentable/sous/util/logging/messages"
 	"github.com/pkg/errors"
 	"github.com/samsalisbury/semv"
 )
@@ -76,16 +77,16 @@ func (ds Deployments) PutbackManifests(defs Defs, olds Manifests) (Manifests, er
 				continue
 			}
 			if string(clusterVal) == v {
-				logging.Log.Debug.Printf("Redundant environment definition: %s=%s", k, v)
+				messages.ReportLogFieldsMessage("Redundant environment definition", logging.DebugLevel, logging.Log, k, v)
 				if was && hadSpec {
 					if _, present := oldSpec.Env[k]; present {
-						logging.Log.Debug.Printf("Env pair %s=%s present in existing manifest: retained.", k, v)
+						messages.ReportLogFieldsMessage("Env pair present in existing manifest: retained", logging.DebugLevel, logging.Log, k, v)
 					} else {
-						logging.Log.Debug.Printf("Env pair %s=%s absent in existing manifest: elided.", k, v)
+						messages.ReportLogFieldsMessage("Env pair absent in existing manifest: deleted", logging.DebugLevel, logging.Log, k, v)
 						delete(spec.Env, k)
 					}
 				} else {
-					logging.Log.Debug.Printf("Manifest for %v or cluster %q absent in existing manifest list: eliding %s=%s.", mid, d.ClusterName, k, v)
+					messages.ReportLogFieldsMessage("Manifest or cluster absent in existing manifest list: deleted", logging.DebugLevel, logging.Log, mid, d.ClusterName, k, v)
 					delete(spec.Env, k)
 				}
 			}
@@ -142,7 +143,7 @@ func (ds Deployments) RawManifests(defs Defs) (Manifests, error) {
 				continue
 			}
 			if string(clusterVal) == v {
-				logging.Log.Debug.Printf("Redundant environment definition: %s=%s", k, v)
+				messages.ReportLogFieldsMessage("Redundant environment definition", logging.DebugLevel, logging.Log, k, v)
 			}
 		}
 		m.Deployments[d.ClusterName] = spec