+++ test/server_integration_test.go
@@ -217,6 +217,7 @@ func (suite integrationServerTests) TestPUTSingleDeployment() {
 		"repo":    "github.com/opentable/sous",
 		"offset":  "",
 		"tag":     "1.0.1",
+		"force":   "false",
 	}
 	data := server.SingleDeploymentBody{}
 	rez, err = suite.client.Retrieve("/single-deployment", params, &data, nil)